feat: allow ci build id and project id as paramenters for the gha and send it to cancel API
